64 bit build support
Tizen toolchain team in Samsung company is planning to build Tizen
for new arm 64 bit architecture - Aarch64. But package "libscl-ui"
cannot be built for this architecture because of build errors.

The main problem is pointer size change, so fix is just cutting lowest 32 bytes
on 64-bit platforms to use them as id for timer.

Change-Id: I8f75a659ac1eb34da2c50e42c6c7f34595366a15
Signed-off-by: Vyacheslav Barinov <v.barinov@samsung.com>